#1626 - Improve testability of AnnotationServiceImpl
- Refactor AnnotationServiceImpl to constructor injection
- Upgrade tests accordingly
- Remove most circular dependencies between layer supports and AnnotationService
- Move project initializers from AnnotationService to ProjectService
