+++ webanno-agreement/src/test/java/de/tudarmstadt/ukp/clarin/webanno/agreement/measures/AgreementMeasureTestSuite_ImplBase.java
@@ -90,12 +90,9 @@ public class AgreementMeasureTestSuite_ImplBase
         layerBehaviorRegistry.init();
             
         LayerSupportRegistryImpl layerRegistry = new LayerSupportRegistryImpl(asList(
-                new SpanLayerSupport(featureSupportRegistry, null, annotationService,
-                    layerBehaviorRegistry),
-                new RelationLayerSupport(featureSupportRegistry, null, annotationService,
-                    layerBehaviorRegistry),
-                new ChainLayerSupport(featureSupportRegistry, null, annotationService,
-                    layerBehaviorRegistry)));
+                new SpanLayerSupport(featureSupportRegistry, null, layerBehaviorRegistry),
+                new RelationLayerSupport(featureSupportRegistry, null, layerBehaviorRegistry),
+                new ChainLayerSupport(featureSupportRegistry, null, layerBehaviorRegistry)));
         layerRegistry.init();
                 
         when(annotationService.listAnnotationLayer(any())).thenReturn(layers);
@@ -103,7 +100,8 @@ public class AgreementMeasureTestSuite_ImplBase
             .thenReturn(features);
         when(annotationService.getAdapter(any(AnnotationLayer.class))).then(_call -> {
             AnnotationLayer l = _call.getArgument(0);
-            return layerRegistry.getLayerSupport(l).createAdapter(l);
+            return layerRegistry.getLayerSupport(l).createAdapter(l, 
+                () -> annotationService.listAnnotationFeature(l));
         });
     }
