+++ webanno-api-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/annotation/layer/SpanLayerSupport.java
@@ -20,7 +20,9 @@ package de.tudarmstadt.ukp.clarin.webanno.api.annotation.layer;
 import static java.util.Arrays.asList;
 import static java.util.stream.Collectors.toList;
 
+import java.util.Collection;
 import java.util.List;
+import java.util.function.Supplier;
 
 import org.apache.uima.cas.CAS;
 import org.apache.uima.resource.metadata.TypeDescription;
@@ -30,10 +32,10 @@ import org.apache.wicket.model.IModel;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.InitializingBean;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 
-import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationSchemaService;
 import de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.adapter.SpanAdapter;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.adapter.SpanLayerBehavior;
@@ -50,19 +52,18 @@ public class SpanLayerSupport
     private final Logger log = LoggerFactory.getLogger(getClass());
     
     private final ApplicationEventPublisher eventPublisher;
-    private final AnnotationSchemaService schemaService;
     private final LayerBehaviorRegistry layerBehaviorsRegistry;
     
     private String layerSupportId;
     private List<LayerType> types;
 
+    @Autowired
     public SpanLayerSupport(FeatureSupportRegistry aFeatureSupportRegistry,
-            ApplicationEventPublisher aEventPublisher, AnnotationSchemaService aSchemaService,
+            ApplicationEventPublisher aEventPublisher,
             LayerBehaviorRegistry aLayerBehaviorsRegistry)
     {
         super(aFeatureSupportRegistry);
         eventPublisher = aEventPublisher;
-        schemaService = aSchemaService;
         layerBehaviorsRegistry = aLayerBehaviorsRegistry;
     }
 
@@ -97,10 +98,11 @@ public class SpanLayerSupport
     }
     
     @Override
-    public SpanAdapter createAdapter(AnnotationLayer aLayer)
+    public SpanAdapter createAdapter(AnnotationLayer aLayer,
+            Supplier<Collection<AnnotationFeature>> aFeatures)
     {
         SpanAdapter adapter = new SpanAdapter(getLayerSupportRegistry(), featureSupportRegistry,
-                eventPublisher, aLayer, () -> schemaService.listAnnotationFeature(aLayer),
+                eventPublisher, aLayer, aFeatures,
                 layerBehaviorsRegistry.getLayerBehaviors(this, SpanLayerBehavior.class));
         
         return adapter;
@@ -120,9 +122,10 @@ public class SpanLayerSupport
     }
     
     @Override
-    public SpanRenderer getRenderer(AnnotationLayer aLayer)
+    public SpanRenderer createRenderer(AnnotationLayer aLayer,
+            Supplier<Collection<AnnotationFeature>> aFeatures)
     {
-        return new SpanRenderer(createAdapter(aLayer), getLayerSupportRegistry(),
+        return new SpanRenderer(createAdapter(aLayer, aFeatures), getLayerSupportRegistry(),
                 featureSupportRegistry,
                 layerBehaviorsRegistry.getLayerBehaviors(this, SpanLayerBehavior.class));
     }
