+++ webanno-project/src/main/java/de/tudarmstadt/ukp/clarin/webanno/project/initializers/LemmaLayerInitializer.java
@@ -15,33 +15,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.api.dao.initializers;
+package de.tudarmstadt.ukp.clarin.webanno.project.initializers;
 
 import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.SPAN_TYPE;
 import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.SINGLE_TOKEN;
 import static de.tudarmstadt.ukp.clarin.webanno.model.OverlapMode.NO_OVERLAP;
-import static de.tudarmstadt.ukp.clarin.webanno.model.ValidationMode.NEVER;
+import static java.util.Arrays.asList;
 
 import java.io.IOException;
-import java.util.Collections;
 import java.util.List;
 
+import org.apache.uima.cas.CAS;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationSchemaService;
+import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationFeature;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
+import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Lemma;
 import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token;
 
 @Component
-public class TokenLayerInitializer
+public class LemmaLayerInitializer
     implements LayerInitializer
 {
     private final AnnotationSchemaService annotationSchemaService;
-
+    
     @Autowired
-    public TokenLayerInitializer(AnnotationSchemaService aAnnotationSchemaService)
+    public LemmaLayerInitializer(AnnotationSchemaService aAnnotationSchemaService)
     {
         annotationSchemaService = aAnnotationSchemaService;
     }
@@ -49,20 +51,33 @@ public class TokenLayerInitializer
     @Override
     public List<Class<? extends ProjectInitializer>> getDependencies()
     {
-        return Collections.emptyList();
+        // Because locks to token boundaries and attaches to token
+        return asList(TokenLayerInitializer.class);
     }
 
     @Override
     public void configure(Project aProject) throws IOException
     {
-        AnnotationLayer tokenLayer = new AnnotationLayer(Token.class.getName(), "Token", SPAN_TYPE,
-                aProject, true, SINGLE_TOKEN, NO_OVERLAP);
-        
-        // Since the user cannot turn off validation for the token layer if there is any kind of
-        // problem with the validation functionality we are conservative here and disable validation
-        // from the start.
-        tokenLayer.setValidationMode(NEVER);
+        AnnotationLayer tokenLayer = annotationSchemaService.findLayer(aProject,
+                Token.class.getName());
 
-        annotationSchemaService.createLayer(tokenLayer);
+        AnnotationFeature tokenLemmaFeature = new AnnotationFeature(aProject, tokenLayer, "lemma",
+                "lemma", Lemma.class.getName());
+        annotationSchemaService.createFeature(tokenLemmaFeature);
+
+        AnnotationLayer lemmaLayer = new AnnotationLayer(Lemma.class.getName(), "Lemma", SPAN_TYPE,
+                aProject, true, SINGLE_TOKEN, NO_OVERLAP);
+        lemmaLayer.setAttachType(tokenLayer);
+        lemmaLayer.setAttachFeature(tokenLemmaFeature);
+        annotationSchemaService.createLayer(lemmaLayer);
+
+        AnnotationFeature lemmaFeature = new AnnotationFeature();
+        lemmaFeature.setDescription("lemma Annotation");
+        lemmaFeature.setName("value");
+        lemmaFeature.setType(CAS.TYPE_NAME_STRING);
+        lemmaFeature.setProject(aProject);
+        lemmaFeature.setUiName("Lemma");
+        lemmaFeature.setLayer(lemmaLayer);
+        annotationSchemaService.createFeature(lemmaFeature);
     }
 }
