+++ webanno-project/src/main/java/de/tudarmstadt/ukp/clarin/webanno/project/initializers/MorphologicalFeaturesLayerInitializer.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.api.dao.initializers;
+package de.tudarmstadt.ukp.clarin.webanno.project.initializers;
 
 import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.SPAN_TYPE;
 import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.SINGLE_TOKEN;
@@ -27,26 +27,24 @@ import java.util.List;
 
 import org.apache.uima.cas.CAS;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.stereotype.Component;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationSchemaService;
-import de.tudarmstadt.ukp.clarin.webanno.api.dao.JsonImportUtil;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationFeature;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
-import de.tudarmstadt.ukp.clarin.webanno.model.TagSet;
-import de.tudarmstadt.ukp.dkpro.core.api.lexmorph.type.pos.POS;
+import de.tudarmstadt.ukp.dkpro.core.api.lexmorph.type.morph.MorphologicalFeatures;
+import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Lemma;
 import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token;
 
 @Component
-public class PartOfSpeechLayerInitializer
+public class MorphologicalFeaturesLayerInitializer
     implements LayerInitializer
 {
     private final AnnotationSchemaService annotationSchemaService;
 
     @Autowired
-    public PartOfSpeechLayerInitializer(AnnotationSchemaService aAnnotationSchemaService)
+    public MorphologicalFeaturesLayerInitializer(AnnotationSchemaService aAnnotationSchemaService)
     {
         annotationSchemaService = aAnnotationSchemaService;
     }
@@ -61,34 +59,28 @@ public class PartOfSpeechLayerInitializer
     @Override
     public void configure(Project aProject) throws IOException
     {
-        TagSet posTagSet = JsonImportUtil.importTagSetFromJson(aProject,
-                new ClassPathResource("/tagsets/mul-pos-ud2.json").getInputStream(),
-                annotationSchemaService);
-
         AnnotationLayer tokenLayer = annotationSchemaService.findLayer(aProject,
                 Token.class.getName());
 
-        AnnotationLayer posLayer = new AnnotationLayer(POS.class.getName(), "Part of speech",
-                SPAN_TYPE, aProject, true, SINGLE_TOKEN, NO_OVERLAP);
+        AnnotationFeature tokenMorphFeature = new AnnotationFeature(aProject, tokenLayer, "morph",
+                "morph", Lemma.class.getName());
+        annotationSchemaService.createFeature(tokenMorphFeature);
 
-        AnnotationFeature tokenPosFeature = new AnnotationFeature(aProject, tokenLayer, "pos",
-                "pos", POS.class.getName());
-        annotationSchemaService.createFeature(tokenPosFeature);
+        AnnotationLayer morphLayer = new AnnotationLayer(MorphologicalFeatures.class.getName(),
+                "Morphological features", SPAN_TYPE, aProject, true, SINGLE_TOKEN, NO_OVERLAP);
+        morphLayer.setAttachType(tokenLayer);
+        morphLayer.setAttachFeature(tokenMorphFeature);
+        annotationSchemaService.createLayer(morphLayer);
 
-        posLayer.setAttachType(tokenLayer);
-        posLayer.setAttachFeature(tokenPosFeature);
-        annotationSchemaService.createLayer(posLayer);
-
-        AnnotationFeature xpos = new AnnotationFeature(aProject, posLayer, "PosValue",
-                "XPOS", CAS.TYPE_NAME_STRING, "XPOS", null);
-        xpos.setDescription("Language-specific part-of-speech tag");
-        annotationSchemaService.createFeature(xpos);
-
-        AnnotationFeature upos = new AnnotationFeature(aProject, posLayer, "coarseValue",
-                "UPOS", CAS.TYPE_NAME_STRING, "UPOS", posTagSet);
-        upos.setDescription("Universal part-of-speech tag");
-        annotationSchemaService.createFeature(upos);
-
-                
+        AnnotationFeature valueFeature = new AnnotationFeature();
+        valueFeature.setDescription("Morphological features");
+        valueFeature.setName("value");
+        valueFeature.setType(CAS.TYPE_NAME_STRING);
+        valueFeature.setProject(aProject);
+        valueFeature.setUiName("Features");
+        valueFeature.setLayer(morphLayer);
+        valueFeature.setIncludeInHover(true);
+        valueFeature.setVisible(false);
+        annotationSchemaService.createFeature(valueFeature);
     }
 }
