+++ webanno-project/src/main/java/de/tudarmstadt/ukp/clarin/webanno/project/initializers/TokenLayerInitializer.java
@@ -15,7 +15,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.api.dao.initializers;
+package de.tudarmstadt.ukp.clarin.webanno.project.initializers;
+
+import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.SPAN_TYPE;
+import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.SINGLE_TOKEN;
+import static de.tudarmstadt.ukp.clarin.webanno.model.OverlapMode.NO_OVERLAP;
+import static de.tudarmstadt.ukp.clarin.webanno.model.ValidationMode.NEVER;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -25,16 +30,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationSchemaService;
+import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
+import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token;
 
 @Component
-public class DefaultTagSetsInitializer
-    implements TagSetInitializer
+public class TokenLayerInitializer
+    implements LayerInitializer
 {
     private final AnnotationSchemaService annotationSchemaService;
-    
+
     @Autowired
-    public DefaultTagSetsInitializer(AnnotationSchemaService aAnnotationSchemaService)
+    public TokenLayerInitializer(AnnotationSchemaService aAnnotationSchemaService)
     {
         annotationSchemaService = aAnnotationSchemaService;
     }
@@ -48,20 +55,14 @@ public class DefaultTagSetsInitializer
     @Override
     public void configure(Project aProject) throws IOException
     {
-//        JsonImportUtil.importTagSetFromJson(aProject,
-//                new ClassPathResource("/tagsets/de-pos-stts.json").getInputStream(),
-//                annotationSchemaService);
-//        JsonImportUtil.importTagSetFromJson(aProject,
-//                new ClassPathResource("/tagsets/de-dep-tiger.json").getInputStream(),
-//                annotationSchemaService);
-//        JsonImportUtil.importTagSetFromJson(aProject,
-//                new ClassPathResource("/tagsets/en-dep-sd.json").getInputStream(),
-//                annotationSchemaService);
-//        JsonImportUtil.importTagSetFromJson(aProject,
-//                new ClassPathResource("/tagsets/en-pos-ptb-tt.json").getInputStream(),
-//                annotationSchemaService);
-//        JsonImportUtil.importTagSetFromJson(aProject,
-//                new ClassPathResource("/tagsets/mul-pos-upos.json").getInputStream(),
-//                annotationSchemaService);
+        AnnotationLayer tokenLayer = new AnnotationLayer(Token.class.getName(), "Token", SPAN_TYPE,
+                aProject, true, SINGLE_TOKEN, NO_OVERLAP);
+        
+        // Since the user cannot turn off validation for the token layer if there is any kind of
+        // problem with the validation functionality we are conservative here and disable validation
+        // from the start.
+        tokenLayer.setValidationMode(NEVER);
+
+        annotationSchemaService.createLayer(tokenLayer);
     }
 }
