+++ webanno-remote/pom.xml
@@ -195,6 +195,11 @@
       <scope>test</scope>
     </dependency>
     <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-test-autoconfigure</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
       <groupId>org.apache.tomcat.embed</groupId>
       <artifactId>tomcat-embed-core</artifactId>
       <scope>test</scope>
@@ -214,11 +219,6 @@
           <artifactId>spring-boot-starter-logging</artifactId>
         </exclusion>
       </exclusions>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-data-jpa</artifactId>
-      <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.springframework.security</groupId>
@@ -250,6 +250,17 @@
       <artifactId>webanno-io-text</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-data-jpa</artifactId>
+      <scope>test</scope>
+      <exclusions>
+        <exclusion>
+          <groupId>ch.qos.logback</groupId>
+          <artifactId>logback-classic</artifactId>
+        </exclusion>
+      </exclusions>
+    </dependency>
   </dependencies>
   <build>
     <pluginManagement>
@@ -268,6 +279,11 @@
               <ignoredDependency>org.springframework.boot:spring-boot-starter-data-jpa</ignoredDependency>
               <ignoredDependency>com.jayway.jsonpath:json-path</ignoredDependency>
               <ignoredDependency>com.h2database:h2</ignoredDependency>
+              <ignoredDependency>de.tudarmstadt.ukp.clarin.webanno:webanno-io-text</ignoredDependency>
+              <ignoredDependency>de.tudarmstadt.ukp.clarin.webanno:webanno-project</ignoredDependency>
+              <ignoredDependency>de.tudarmstadt.ukp.clarin.webanno:webanno-api-dao</ignoredDependency>
+              <ignoredDependency>org.springframework.boot:spring-boot-test-autoconfigure</ignoredDependency>
+              <ignoredDependency>org.springframework.security:spring-security-config</ignoredDependency>
             </ignoredDependencies>
             <usedDependencies combine.children="append">
               <!--
