--- webanno-remote/src/main/resources/RemoteApiController2Test.properties
@@ -1,1 +1,0 @@
-repository.path=target/RemoteApiController2Test/repository
