+++ webanno-ui-project/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/project/detail/ProjectDetailPanel.java
@@ -157,7 +157,7 @@ public class ProjectDetailPanel
                 projectService.createProjectPermission(
                         new ProjectPermission(project, username, PermissionLevel.ANNOTATOR));
 
-                annotationService.initializeProject(project);
+                projectService.initializeProject(project);
             }
             catch (IOException e) {
                 error("Project repository path not found " + ":"
